refactor(web,ui): 一覧ポップオーバーのピル語彙統一・低評価UI撤去・per-user hook統合 (SPR-257 PR①)
一覧刷新の土台整理。3ビュー実装（PR②）の前に触る場所を整える。

- audio-button.tsx（660行）からポップオーバーを audio-button-popover.tsx へ分離。
  アクション層は詳細ページ・モーダルと同じ ActionPillRow（sm）+ YouTube/詳細ピルに統一
- 低評価（👎）UI を全経路から撤去（製品判断確定済み・データ/action は温存）。
  isDisliked/onDislikeToggle props と initialIsDisliked の受け渡しチェーンを削除
- per-user 状態の正本を hooks/use-audio-button-engagement.ts に統合
  （一覧の一括初期値注入と詳細/モーダルの自己取得の両対応・楽観的更新+ロールバック・
  未ログインは toast のみに統一）。favorite-client 228行→64行、hero-state は再生系のみに
- ActionPillRow の shareUrl を optional 化（未指定時は共有ピル非表示）
- 死蔵コード削除: like-dislike-buttons / favorite-button / audio-button-card /
  audio-button-list（いずれも消費者ゼロ・SPR-256 撤去の残骸）
- showFavorite prop 撤去（全呼び出しが true 固定の死んだ設定）

検証: pnpm verify / Storybook(a11y含む352件) 通過。Emulator + 実ブラウザで
ポップオーバーのピル構成・低評価不在・用途タグ表示を確認済み。
