+++ packages/ui/src/components/custom/audio-button.tsx
@@ -1,10 +1,7 @@
 /**
- * プール化対応 AudioButton コンポーネント
- *
- * 変更点:
- * - SimpleAudioButton → AudioButton（命名統一）
- * - AudioOnlyPlayer → AudioPlayer（プール化対応）
- * - パフォーマンス最適化
+ * プール化対応 AudioButton コンポーネント（一覧・カルーセル用のコンパクトピル）。
+ * ポップオーバー内容は audio-button-popover.tsx（SPR-257 で分離・ピル語彙化）、
+ * 再生状態機械は use-audio-playback.ts（SPR-258・PlayHero と共有）が正本。
  */
 
 "use client";
@@ -12,28 +9,13 @@
 import type { AudioButton as AudioButtonType } from "@suzumina.click/shared-types";
 import { Popover, PopoverContent, PopoverTrigger } from "@suzumina.click/ui/components/ui/popover";
 import { cn } from "@suzumina.click/ui/lib/utils";
-import {
-	ArrowRight,
-	Clock,
-	Heart,
-	Loader2,
-	Lock,
-	MoreHorizontal,
-	Pause,
-	Play,
-	ThumbsDown,
-	ThumbsUp,
-	User,
-	Video,
-} from "lucide-react";
-import { useCallback, useId, useState } from "react";
+import { Loader2, MoreHorizontal, Pause, Play } from "lucide-react";
+import { useCallback, useState } from "react";
 import { PROGRESS_TICK_MS } from "../../lib/playback-constants";
+import { AudioButtonPopoverContent } from "./audio-button-popover";
 import { AudioPlayer } from "./audio-player";
 import { HighlightText } from "./highlight-text";
-import { TagList } from "./tag-list";
 import { useAudioPlayback } from "./use-audio-playback";
-import { XIcon } from "./x-icon";
-import { YoutubeIcon } from "./youtube-icon";
 
 interface AudioButtonProps {
 	audioButton: AudioButtonType;
@@ -53,10 +35,6 @@ interface AudioButtonProps {
 	isLiked?: boolean;
 	/** onFavoriteToggle と同様、未ログイン時のガードは呼び出し元の責務 */
 	onLikeToggle?: () => void;
-	// 低評価関連
-	isDisliked?: boolean;
-	/** onFavoriteToggle と同様、未ログイン時のガードは呼び出し元の責務 */
-	onDislikeToggle?: () => void;
 	// ハイライト関連
 	searchQuery?: string;
 	highlightClassName?: string;
@@ -64,401 +42,6 @@ interface AudioButtonProps {
 	isAuthenticated?: boolean;
 	/** X共有の intent URL。指定時のみポップオーバーに「共有」アクションを表示（組み立ては呼び出し元の責務） */
 	xShareUrl?: string;
-}
-
-interface AudioButtonPopoverContentProps {
-	audioButton: AudioButtonType;
-	duration: number;
-	youtubeUrl: string;
-	isFavorite: boolean;
-	onFavoriteToggle?: () => void;
-	isLiked: boolean;
-	onLikeToggle?: () => void;
-	isDisliked: boolean;
-	onDislikeToggle?: () => void;
-	showDetailLink: boolean;
-	onDetailClick?: () => void;
-	onPopoverClose: () => void;
-	searchQuery?: string;
-	highlightClassName?: string;
-	isAuthenticated: boolean;
-	xShareUrl?: string;
-}
-
-// Extracted components for reducing complexity
-function PopoverTitle({
-	text,
-	searchQuery,
-	highlightClassName,
-}: {
-	text: string;
-	searchQuery?: string;
-	highlightClassName?: string;
-}) {
-	if (!searchQuery) return <>{text}</>;
-	return (
-		<HighlightText
-			text={text}
-			searchQuery={searchQuery}
-			highlightClassName={highlightClassName || "bg-primary/20 text-foreground px-1 rounded"}
-		/>
-	);
-}
-
-function PopoverDescription({
-	description,
-	searchQuery,
-	highlightClassName,
-}: {
-	description: string | undefined;
-	searchQuery?: string;
-	highlightClassName?: string;
-}) {
-	if (!description) return null;
-	return (
-		<p className="text-sm text-muted-foreground mt-1 leading-relaxed">
-			<PopoverTitle
-				text={description}
-				searchQuery={searchQuery}
-				highlightClassName={highlightClassName}
-			/>
-		</p>
-	);
-}
-
-// 見出し層: タイトル + 秒数/再生数チップ + 説明
-function PopoverHeader({
-	audioButton,
-	duration,
-	searchQuery,
-	highlightClassName,
-}: {
-	audioButton: AudioButtonType;
-	duration: number;
-	searchQuery?: string;
-	highlightClassName?: string;
-}) {
-	return (
-		<div>
-			<h4 className="font-semibold text-base text-foreground leading-tight">
-				<PopoverTitle
-					text={audioButton.buttonText}
-					searchQuery={searchQuery}
-					highlightClassName={highlightClassName}
-				/>
-			</h4>
-			<div className="mt-1.5 flex items-center gap-2 text-xs text-muted-foreground">
-				<span className="inline-flex items-center gap-1 rounded-full bg-muted px-2 py-0.5 font-semibold text-foreground">
-					<Clock className="h-3 w-3" />
-					{duration.toFixed(1)}秒
-				</span>
-				<span>再生 {audioButton.stats.playCount}回</span>
-			</div>
-			<PopoverDescription
-				description={audioButton.description}
-				searchQuery={searchQuery}
-				highlightClassName={highlightClassName}
-			/>
-		</div>
-	);
-}
-
-// 出典リスト層: 作成者・元動画への導線
-function PopoverMetaList({ audioButton }: { audioButton: AudioButtonType }) {
-	return (
-		<div className="overflow-hidden rounded-lg border border-border">
-			<a
-				href={`/users/${audioButton.creatorId}`}
-				className="flex min-h-[38px] items-center gap-2 px-3 py-2 text-xs hover:bg-accent transition-colors"
-				onClick={(e) => e.stopPropagation()}
-			>
-				<User className="h-3.5 w-3.5 flex-none text-muted-foreground" />
-				<span className="min-w-0 flex-1 truncate">{audioButton.creatorName}</span>
-				<span className="flex-none text-muted-foreground">作成者</span>
-			</a>
-			{audioButton.videoTitle && (
-				<a
-					href={`/videos/${audioButton.videoId}`}
-					className="flex min-h-[38px] items-center gap-2 border-t border-border px-3 py-2 text-xs hover:bg-accent transition-colors"
-					onClick={(e: React.MouseEvent) => e.stopPropagation()}
-					title={audioButton.videoTitle}
-				>
-					<Video className="h-3.5 w-3.5 flex-none text-muted-foreground" />
-					<span className="min-w-0 flex-1 truncate">{audioButton.videoTitle}</span>
-					<span className="flex-none text-muted-foreground">元動画</span>
-				</a>
-			)}
-		</div>
-	);
-}
-
-// Tags section component
-function PopoverTags({
-	audioButton,
-	searchQuery,
-	highlightClassName,
-}: {
-	audioButton: AudioButtonType;
-	searchQuery?: string;
-	highlightClassName?: string;
-}) {
-	if (!audioButton.tags || audioButton.tags.length === 0) return null;
-	return (
-		<div>
-			<p className="text-xs text-muted-foreground mb-2">タグ</p>
-			<TagList
-				tags={audioButton.tags}
-				variant="outline"
-				showIcon={true}
-				searchQuery={searchQuery}
-				highlightClassName={highlightClassName || "bg-primary/20 text-foreground px-1 rounded"}
-				size="default"
-			/>
-		</div>
-	);
-}
-
-// Favorite button component
-function FavoriteButton({
-	isFavorite,
-	onFavoriteToggle,
-	describedBy,
-}: {
-	isFavorite?: boolean;
-	onFavoriteToggle?: () => void;
-	describedBy?: string;
-}) {
-	if (!onFavoriteToggle) return null;
-
-	return (
-		<button
-			type="button"
-			onClick={(e) => {
-				e.stopPropagation();
-				onFavoriteToggle();
-			}}
-			aria-label={isFavorite ? "お気に入りを解除" : "お気に入りに追加"}
-			aria-describedby={describedBy}
-			className={cn(
-				"flex items-center justify-center gap-1.5 h-10 px-2.5 rounded-lg text-xs font-semibold transition-colors hover:bg-accent",
-				isFavorite ? "text-suzuka-600" : "text-muted-foreground hover:text-foreground",
-			)}
-		>
-			<Heart className={cn("h-4 w-4", isFavorite && "fill-current")} />
-		</button>
-	);
-}
-
-// Like/Dislike buttons component
-function LikeDislikeButtons({
-	audioButton,
-	isLiked,
-	onLikeToggle,
-	isDisliked,
-	onDislikeToggle,
-	describedBy,
-}: {
-	audioButton: AudioButtonType;
-	isLiked?: boolean;
-	onLikeToggle?: () => void;
-	isDisliked?: boolean;
-	onDislikeToggle?: () => void;
-	describedBy?: string;
-}) {
-	if (!onLikeToggle) return null;
-
-	return (
-		<div className="flex items-center">
-			<button
-				type="button"
-				onClick={(e) => {
-					e.stopPropagation();
-					onLikeToggle();
-				}}
-				aria-describedby={describedBy}
-				className={cn(
-					"flex items-center gap-1.5 h-10 px-2.5 rounded-lg text-xs font-semibold transition-colors hover:bg-accent",
-					isLiked ? "text-suzuka-600" : "text-muted-foreground hover:text-foreground",
-				)}
-			>
-				<ThumbsUp className={cn("h-4 w-4", isLiked && "fill-current")} />
-				<span>{audioButton.stats.likeCount}</span>
-			</button>
-			{onDislikeToggle && (
-				<button
-					type="button"
-					onClick={(e) => {
-						e.stopPropagation();
-						onDislikeToggle();
-					}}
-					aria-label={isDisliked ? "低評価を取り消す" : "低評価する"}
-					aria-describedby={describedBy}
-					className="flex items-center justify-center h-10 px-2.5 rounded-lg text-xs font-semibold text-muted-foreground transition-colors hover:bg-accent hover:text-foreground"
-				>
-					<ThumbsDown className={cn("h-4 w-4", isDisliked && "fill-current")} />
-				</button>
-			)}
-		</div>
-	);
-}
-
-// Action buttons component
-function PopoverActions({
-	audioButton,
-	youtubeUrl,
-	isFavorite,
-	onFavoriteToggle,
-	isLiked,
-	onLikeToggle,
-	isDisliked,
-	onDislikeToggle,
-	showDetailLink,
-	onDetailClick,
-	onPopoverClose,
-	describedBy,
-	xShareUrl,
-}: {
-	audioButton: AudioButtonType;
-	youtubeUrl: string;
-	isFavorite?: boolean;
-	onFavoriteToggle?: () => void;
-	isLiked?: boolean;
-	onLikeToggle?: () => void;
-	isDisliked?: boolean;
-	onDislikeToggle?: () => void;
-	showDetailLink?: boolean;
-	onDetailClick?: () => void;
-	onPopoverClose?: () => void;
-	describedBy?: string;
-	xShareUrl?: string;
-}) {
-	return (
-		<div className="flex gap-1 items-center flex-wrap">
-			<FavoriteButton
-				isFavorite={isFavorite}
-				onFavoriteToggle={onFavoriteToggle}
-				describedBy={describedBy}
-			/>
-
-			<LikeDislikeButtons
-				audioButton={audioButton}
-				isLiked={isLiked}
-				onLikeToggle={onLikeToggle}
-				isDisliked={isDisliked}
-				onDislikeToggle={onDislikeToggle}
-				describedBy={describedBy}
-			/>
-
-			<a
-				href={youtubeUrl}
-				target="_blank"
-				rel="noopener noreferrer"
-				className="flex items-center gap-1.5 h-10 px-2.5 rounded-lg text-xs font-semibold bg-destructive/10 text-destructive hover:bg-destructive/20 transition-colors"
-				onClick={(e) => e.stopPropagation()}
-			>
-				<YoutubeIcon className="h-4 w-4" />
-				YouTube
-			</a>
-
-			{xShareUrl && (
-				<a
-					href={xShareUrl}
-					target="_blank"
-					rel="noopener noreferrer"
-					className="flex items-center gap-1.5 h-10 px-2.5 rounded-lg text-xs font-semibold bg-foreground/5 text-foreground hover:bg-foreground/10 transition-colors"
-					onClick={(e) => e.stopPropagation()}
-					aria-label={`「${audioButton.buttonText}」をXで共有`}
-				>
-					<XIcon className="h-3.5 w-3.5" />
-					共有
-				</a>
-			)}
-
-			{showDetailLink && onDetailClick && (
-				<button
-					type="button"
-					onClick={(e) => {
-						e.stopPropagation();
-						onDetailClick();
-						onPopoverClose?.();
-					}}
-					aria-label="詳細ページを開く"
-					className="flex items-center gap-1 h-10 px-2.5 ml-auto rounded-lg text-xs font-semibold text-muted-foreground hover:text-foreground transition-colors whitespace-nowrap"
-				>
-					詳細
-					<ArrowRight className="h-3 w-3" />
-				</button>
-			)}
-		</div>
-	);
-}
-
-function AudioButtonPopoverContent({
-	audioButton,
-	duration,
-	youtubeUrl,
-	isFavorite,
-	onFavoriteToggle,
-	isLiked,
-	onLikeToggle,
-	isDisliked,
-	onDislikeToggle,
-	showDetailLink,
-	onDetailClick,
-	onPopoverClose,
-	searchQuery,
-	highlightClassName,
-	isAuthenticated,
-	xShareUrl,
-}: AudioButtonPopoverContentProps) {
-	const authNoteId = useId();
-
-	return (
-		<div className="p-4 space-y-4">
-			{/* 見出し層: タイトル・秒数/再生数・説明 */}
-			<PopoverHeader
-				audioButton={audioButton}
-				duration={duration}
-				searchQuery={searchQuery}
-				highlightClassName={highlightClassName}
-			/>
-
-			{/* 出典リスト層: 作成者・元動画 */}
-			<PopoverMetaList audioButton={audioButton} />
-
-			{/* タグ層 */}
-			<PopoverTags
-				audioButton={audioButton}
-				searchQuery={searchQuery}
-				highlightClassName={highlightClassName}
-			/>
-
-			{/* アクション層 */}
-			<div className="space-y-2 border-t border-border pt-3">
-				{!isAuthenticated && (
-					<p id={authNoteId} className="flex items-center gap-1.5 text-xs text-muted-foreground">
-						<Lock className="h-3 w-3" />
-						お気に入り・評価にはログインが必要です
-					</p>
-				)}
-				<PopoverActions
-					audioButton={audioButton}
-					youtubeUrl={youtubeUrl}
-					isFavorite={isFavorite}
-					onFavoriteToggle={onFavoriteToggle}
-					isLiked={isLiked}
-					onLikeToggle={onLikeToggle}
-					isDisliked={isDisliked}
-					onDislikeToggle={onDislikeToggle}
-					showDetailLink={showDetailLink}
-					onDetailClick={onDetailClick}
-					onPopoverClose={onPopoverClose}
-					describedBy={isAuthenticated ? undefined : authNoteId}
-					xShareUrl={xShareUrl}
-				/>
-			</div>
-		</div>
-	);
 }
 
 /**
@@ -475,8 +58,6 @@ export function AudioButton({
 	onFavoriteToggle,
 	isLiked = false,
 	onLikeToggle,
-	isDisliked = false,
-	onDislikeToggle,
 	searchQuery,
 	highlightClassName,
 	isAuthenticated = false,
@@ -599,8 +180,6 @@ export function AudioButton({
 						onFavoriteToggle={onFavoriteToggle}
 						isLiked={isLiked}
 						onLikeToggle={onLikeToggle}
-						isDisliked={isDisliked}
-						onDislikeToggle={onDislikeToggle}
 						showDetailLink={showDetailLink}
 						onDetailClick={onDetailClick}
 						onPopoverClose={() => setIsPopoverOpen(false)}
